reorganize upload type selector
